+++ src/render.rs
@@ -7,11 +7,11 @@ use palette::{Gradient, LinSrgb, Pixel};
 /// Convert a 2D array of integers to an RGB image array.
 pub fn image(data: Array2<u8>, max: u8, cmap: &Gradient<LinSrgb>) -> Array3<u8> {
     let mut cols = Array3::<u8>::zeros((data.shape()[0], data.shape()[1], 3));
-    let max_inv = 1.0 / max as f64;
+    let max_inv = 1.0 / max as f32;
     let (width, height) = data.dim();
     for yi in 0..height {
         for xi in 0..width {
-            let x = data[(xi, yi)] as f64 * max_inv;
+            let x = data[(xi, yi)] as f32 * max_inv;
             let col = cmap.get(x as f32);
             let u8s: [u8; 3] = col.into_format().into_raw();
             cols.slice_mut(s![xi, yi, ..]).assign(&arr1(&u8s));
